added reset for goals everyday.
Resets goals everyday.
logic error might exist for saving data: Current code pulls logs from
userlog schema and gets the log with a date on the same day as the
current date. A use case error may occur is a user saves a log right
before the next day triggers. For example, a user logs in at 11:59
monday, triggers a save, and save occurs at 12:01 tuesday. This may not
actually result in an error, but its good to check it.

Also, parse date may be different from device date.